% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/afex-package.R
\docType{package}
\name{afex-package}
\alias{afex}
\alias{afex-package}
\title{\packageTitle{afex}}
\description{
\packageDescription{afex}
}
\details{
The DESCRIPTION file:
\tabular{ll}{
Package: \tab afex\cr
Type: \tab Package\cr
Version: \tab 0.22-0\cr
Date: \tab 2018-08-09\cr
Depends: \tab R (>= 3.1.0), lme4 (>= 1.1-8), emmeans (>= 1.1.2)\cr
Encoding: \tab UTF-8\cr
License: \tab GPL (>=2)\cr
URL: \tab http://afex.singmann.science/, https://github.com/singmann/afex\cr
}
}
\seealso{
Useful links:
\itemize{
  \item \url{http://afex.singmann.science/}
  \item \url{https://github.com/singmann/afex}
  \item Report bugs at \url{https://github.com/singmann/afex/issues}
}

}
\author{
\packageAuthor{afex}

Maintainer: \packageMaintainer{afex}
}
