% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qqnorm.afex_aov.R
\name{qqnorm.afex_aov}
\alias{qqnorm.afex_aov}
\title{Plot Normal Quantile-Quantile Plots for \code{afex_aov} models}
\usage{
\method{qqnorm}{afex_aov}(
  afex_aov,
  type = c("marginal", "univariate", "multivariate"),
  qqbands = TRUE,
  detrend = FALSE,
  return = c("plot", "data"),
  ...
)
}
\arguments{
\item{afex_aov}{\code{afex_aov} object.}

\item{type}{(ignored if model has only between subject effects.) Type of 
error to plot (can be abbreviated):
\describe{
  \item{\code{"marginal}}{Plot the margianl residuals.}
  \item{\code{"univariate}}{Plot the residuals for each error term used in
  the ANOVA table.}
  \item{\code{"multivariate}}{Plot the residuals for each cell in the
  within-subjects effects.}
}}

\item{qqbands}{Should 95% confidence bands be plotted? (requires
\code{qqplotr}).}

\item{detrend}{Should the plot objects be detrended? This may help reducing
visual bias. (requires \code{qqplotr}).}

\item{return}{Should a \code{ggplot} be returned, or a data frame of the
requested errors.}

\item{...}{Not used.}
}
\description{
Plot Normal qq-plots for the residuals / error terms used in an
\code{afex_aov} model. For models with within-subjects effects, qq-plots can
be shown for each error term in the ANOVA table, or for each within-subject
cell (in the multivariate model).
}
\examples{
examples/examples.qqnorm.afex_aov.R

}
\author{
Mattan S. Ben-Shachar
}
