% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/residuals.R
\name{residuals.afex_aov}
\alias{residuals.afex_aov}
\alias{fitted.afex_aov}
\title{Extract Residuals and Fitted Values from \code{afex_aov} objects}
\usage{
\method{residuals}{afex_aov}(object, append = FALSE, colname_residuals = ".residuals", ...)

\method{fitted}{afex_aov}(object, append = FALSE, colname_fitted = ".fitted", ...)
}
\arguments{
\item{object}{\code{afex_aov} object.}

\item{append}{If set to \code{TRUE} returns the residuals/fitted values
appended as an additional column to the long data. Recommended when data
was aggregated across within conditions.}

\item{colname_residuals, colname_fitted}{Name of the appended column when
\code{append = TRUE}.}

\item{...}{Additional arguments passed to
\code{residuals.lm}/\code{fitted.lm}.}
}
\value{
A vector of residuals/fitted values corresponding to the data in
  \code{object$data$long}, or if \code{append = TRUE} a data frame with an
  additional column of residuals/fitted values.
}
\description{
Extract Residuals and Fitted Values from \code{afex_aov} objects.
}
\author{
Mattan S. Ben-Shachar
}
