% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/set_contrasts.R
\name{set_sum_contrasts}
\alias{set_default_contrasts}
\alias{set_deviation_contrasts}
\alias{set_effects_contrasts}
\alias{set_sum_contrasts}
\alias{set_treatment_contrasts}
\title{Set global contrasts}
\usage{
set_sum_contrasts()

set_deviation_contrasts()

set_effects_contrasts()

set_default_contrasts()

set_treatment_contrasts()
}
\value{
nothing. These functions are called for their side effects to change the global options.
}
\description{
These functions are simple wrappers to set contrasts globally via \code{options(contrasts = ...)}.
}
\details{
\code{set_deviation_contrasts} and \code{set_effects_contrasts} are wrappers for \code{set_sum_contrasts}. Likewise, \code{set_default_contrasts} is a wrapper to \code{set_treatment_contrasts()}.
}

