% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{package}
\name{afex-package}
\alias{afex-package}
\title{The afex Package}
\description{
Analysis of Factorial Experiments.
}
\details{
\tabular{ll}{
Package: \tab afex\cr
Type: \tab Package\cr
Version: \tab 0.10-113\cr
Date: \tab 2014-08-03\cr
Depends: \tab R (>= 3.0.0), car, lme4 (>= 1.0.5), pbkrtest (>= 0.3-6), reshape2\cr
Encoding: \tab latin1\cr
License: \tab GPL (>=3)\cr
URL: \tab http://www.psychologie.uni-freiburg.de/Members/singmann/R/afex\cr
}

Provides convenience functions for analyzing factorial experiments using ANOVA or mixed-models. ez.glm() and aov.car() allow convenient calculation of between, within (i.e., repeated-measures), or mixed between-within (i.e., split-plot) ANOVAs for data in the long format (i.e., one observation per row) wrapping car::Anova() (aggregating more then one observation per individual and cell of the design), per default returning a print ready ANOVA table. Function mixed() fits a mixed model using lme4::lmer() and computes p-values for all effects in the model using either the Kenward-Roger approximation of degrees of freedom (LMM only), parametric bootstrap (LMMs and GLMMs) or likelihood ratio tests (LMMs and GLMMs). afex uses type 3 sums of squares as default (imitating commercial statistical software). Furthermore, compare.2.vectors() conveniently compares two vectors using a variety of tests.
}
\author{
Henrik Singmann, with contributions from Ben Bolker, Søren Højsgaard, John Fox, Michael A. Lawrence, Ulf Mertens
}
\keyword{package}

