% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/afex-package.R
\docType{package}
\name{afex-package}
\alias{afex-package}
\title{The afex Package}
\description{
Analysis of Factorial Experiments.
}
\details{
\tabular{ll}{
Package: \tab afex\cr
Type: \tab Package\cr
Version: \tab 0.14-2\cr
Date: \tab 2015-08-17\cr
Depends: \tab R (>= 3.0.0), lme4 (>= 1.0.5), reshape2, lsmeans (>= 2.17)\cr
Encoding: \tab UTF-8\cr
License: \tab GPL (>=3)\cr
URL: \tab https://github.com/singmann/afex\cr
}

Provides convenience functions for analyzing factorial experiments using ANOVA or mixed models. aov_ez(), aov_car(), and aov_4() allow specification of between, within (i.e., repeated-measures), or mixed between-within (i.e., split-plot) ANOVAs for data in long format (i.e., one observation per row), potentially aggregating multiple observations per individual and cell of the design. mixed() fits mixed models using lme4::lmer() and computes p-values for all fixed effects using either Kenward-Roger approximation for degrees of freedom (LMM only), parametric bootstrap (LMMs and GLMMs), or likelihood ratio tests (LMMs and GLMMs). afex uses type 3 sums of squares as default (imitating commercial statistical software).
}
\author{
Henrik Singmann, Ben Bolker, Jake Westfall, with contributions from Søren Højsgaard, John Fox, Michael A. Lawrence, Ulf Mertens
}
\keyword{package}

