% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/afex-package.R
\docType{package}
\name{afex-package}
\alias{afex}
\alias{afex-package}
\title{\packageTitle{afex}}
\description{
\packageDescription{afex}
}
\details{
The DESCRIPTION file:
\packageDESCRIPTION{afex}
}
\seealso{
Useful links:
\itemize{
  \item \url{https://afex.singmann.science/}
  \item \url{https://github.com/singmann/afex}
  \item Report bugs at \url{https://github.com/singmann/afex/issues}
}

}
\author{
\strong{Maintainer}: Henrik Singmann \email{singmann@gmail.com} (\href{https://orcid.org/0000-0002-4842-3657}{ORCID})

Authors:
\itemize{
  \item Ben Bolker
  \item Jake Westfall
  \item Frederik Aust (\href{https://orcid.org/0000-0003-4900-788X}{ORCID})
  \item Mattan S. Ben-Shachar
}

Other contributors:
\itemize{
  \item Søren Højsgaard [contributor]
  \item John Fox [contributor]
  \item Michael A. Lawrence [contributor]
  \item Ulf Mertens [contributor]
  \item Jonathon Love [contributor]
  \item Russell Lenth [contributor]
  \item Rune Haubo Bojesen Christensen [contributor]
}

}
