% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/deprecated.R
\name{aov.car}
\alias{afex-deprecated}
\alias{aov.car}
\alias{ez.glm}
\title{Deprecated functions}
\usage{
aov.car(...)

ez.glm(...)
}
\arguments{
\item{...}{arguments passed from the old functions of the style
\code{foo.bar()} to the new functions \code{foo_bar()}}
}
\description{
These functions have been renamed and deprecated in \pkg{afex}:
\code{aov.car()} (use \code{\link{aov_car}()}),
\code{ez.glm()} (use \code{\link{aov_ez}()}).
}
\keyword{internal}

