% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated.R
\name{test_levene}
\alias{test_levene}
\alias{afex-deprecated}
\alias{test_sphericity}
\title{Deprecated functions}
\usage{
test_levene(...)

test_sphericity(...)
}
\arguments{
\item{...}{arguments passed from the old functions of the style
\code{foo.bar()} to the new functions \code{foo_bar()}}
}
\description{
These functions have been renamed or moved and deprecated in \pkg{afex}:
\code{test_levene()} (use \code{\link[performance]{check_homogeneity}()}),
\code{test_sphericity()} (use \code{\link[performance]{check_sphericity}()}).
}
\keyword{internal}
