% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated.R
\name{aov.car}
\alias{aov.car}
\alias{afex-deprecated}
\alias{ez.glm}
\alias{aov4}
\title{Deprecated functions}
\usage{
aov.car(...)

ez.glm(...)

aov4(...)
}
\arguments{
\item{...}{arguments passed from the old functions of the style
\code{foo.bar()} to the new functions \code{foo_bar()}}
}
\description{
These functions have been renamed and deprecated in \pkg{afex}:
\code{aov.car()} (use \code{\link{aov_car}()}), 
\code{ez.glm()} (use \code{\link{aov_ez}()}),
\code{aov4()} (use \code{\link{aov_4}()}).
}
\keyword{internal}
