% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/residuals.R
\name{residuals.afex_aov}
\alias{residuals.afex_aov}
\title{Extract \code{afex_aov} Model Residuals}
\usage{
\method{residuals}{afex_aov}(object, append = FALSE, ...)
}
\arguments{
\item{object}{\code{afex_aov} object.}

\item{append}{If set to \code{TRUE} returns the residuals appended as an additional column \code{.residuals.} to the long data. 
Recomended when data was aggragated across within conditions.}

\item{...}{Not used.}
}
\value{
A vector of residualts corresponding to the data in \code{object$data$long}, 
or if \code{append = TRUE} a data frame with an additional column \code{.residuals.}.
}
\description{
Extract residuals from \code{afex_aov} objects.
}
\examples{
### Setup ANOVAs
data(obk.long, package = "afex")
between <- aov_car(value ~ treatment*gender + Error(id), data = obk.long)
within <- aov_car(value ~ 1 + Error(id/(phase*hour)), data = obk.long)
mixed <- aov_car(value ~ treatment * gender + Error(id/(phase*hour)), data = obk.long)

# Residuals
residuals(within)
residuals(mixed)
residuals(between)
residuals(between, append = TRUE) # since data was aggragated across within conditions
}
\author{
Mattan S. Ben-Shachar
}
