% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/residuals.R
\name{residuals.afex_aov}
\alias{residuals.afex_aov}
\title{Extract and Plot \code{afex_aov} Model Residuals}
\usage{
\method{residuals}{afex_aov}(object, return_df = FALSE, ...)
}
\arguments{
\item{object}{\code{afex_aov} object.}

\item{return_df}{if \code{TRUE} returns the long data with an additional column \code{residuals}.}

\item{...}{Not used.}
}
\value{
A vector of residualts, or if \code{return_df = TRUE} a data frame with an additional column \code{residuals}.
}
\description{
Extract and Plot \code{afex_aov} Model Residuals
}
\examples{
### Setup ANOVAs
data(obk.long, package = "afex")
between <- aov_car(value ~ treatment*gender + Error(id), data = obk.long)
within <- aov_car(value ~ 1 + Error(id/(phase*hour)), data = obk.long)
mixed <- aov_car(value ~ treatment * gender + Error(id/(phase*hour)), data = obk.long)

# Residuals
residuals(within)
residuals(mixed)
residuals(between)
residuals(between, return_df = TRUE) # since data was aggragated across within conditions
}
\author{
Mattan S. Ben-Shachar
}
