% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/round_ps.R
\encoding{UTF-8}
\name{round_ps}
\alias{round_ps}
\title{Helper function which rounds p-values}
\usage{
round_ps(x)
}
\arguments{
\item{x}{a numeric vector}
}
\value{
A character vector with the same length of x.
}
\description{
p-values are rounded in a sane way: .99 - .01 to two digits, < .01 to three digits, < .001 to four digits.
}
\examples{
round_ps(runif(10))

round_ps(runif(10, 0, .01))

round_ps(runif(10, 0, .001))

round_ps(0.0000000099)

}
\author{
Henrik Singmann
}
